Draw players as divs over table
Creates a div for each player that lives within the table instead of instead of styling individual td's. This allows
us to update the position of single players, rather then redrawing all of them each update. It also allows us to apply
the css styling 'transition' which makes for smooth player movement. It also shows the teleport color when you are standing on it, instead of overwritting that styling.

This also makes the players slightly smaller so that they are not touching the walls of the maze.