Add game rooms
Adds a game room class which manages a group of sockets in a room, and
the logic of driving a game object from player interaction.
This allows for multiple games to take place at the same time.
Rooms have a max number of players, and an amount of time that they
should continue to exist for after the last player leaves.