Don't send down the players id to the player
The player can infer this info from the socket id, and it shouldn't be sent down as
part of the game. This fixes a bug where all other players would think they were
the winning player after a player won the game.

Also fixes a minor whitespace issue with teleport_colors